Get location information from OpenCage

The OpenCage reverse geocoding service is used to obtain all available information for a specific location.

The data from OpenCage should be considered to be the gold standard for address information could and should be used to override information we get elsewhere.

location_getOpenCageInfo(longitude = NULL, latitude = NULL, verbose = FALSE)

Arguments

longitude
Single longitude in decimal degrees E.

latitude
Single latitude in decimal degrees N.

verbose
Logical controlling the generation of progress messages.

Value

Single-row tibble with OpenCage information.

Note

The OpenCage service requires an API key which can be obtained from their web site. This API key must be set as an environment variable with:

Sys.setenv("OPENCAGE_KEY" = "<your api key>")

The OpenCage "free trial" level allows for 1 request/sec and a maximum of 2500 requests per day.

References

https://opencagedata.com

Examples

# \donttest{
library(MazamaLocationUtils)

# Fail gracefully if any resources are not available
try({

  # Wenatchee
  lon <- -120.325278
  lat <- 47.423333

  openCageTbl <- location_getOpenCageInfo(lon, lat)
  dplyr::glimpse(openCageTbl)
  
}, silent = FALSE)

#> Rows: 1
#> Columns: 74
#> $ lat                                         <dbl> 47.42333
#> $ long                                        <dbl> -120.3253
#> $ address                                     <chr> "18 South Miller Street, W…
#> $ confidence                                  <int> 10
#> $ annotations.MGRS                            <chr> "10TGT0176155679"
#> $ annotations.Maidenhead                      <chr> "CN97uk11ao"
#> $ annotations.callingcode                     <int> 1
#> $ annotations.flag                            <chr> "🇺🇸"
#> $ annotations.geohash                         <chr> "c26mvcnhgw4x7s4uqvep"
#> $ annotations.qibla                           <dbl> 19.51
#> $ annotations.DMS.lat                         <chr> "47° 25' 24.00312'' N"
#> $ annotations.DMS.lng                         <chr> "120° 19' 29.49060'' W"
#> $ annotations.FIPS.county                     <chr> "53007"
#> $ annotations.FIPS.state                      <chr> "53"
#> $ annotations.Mercator.x                      <dbl> -13394502
#> $ annotations.Mercator.y                      <dbl> 5979969
#> $ annotations.OSM.edit_url                    <chr> "https://www.openstreetmap…
#> $ annotations.OSM.note_url                    <chr> "https://www.openstreetmap…
#> $ annotations.OSM.url                         <chr> "https://www.openstreetmap…
#> $ annotations.UN_M49.statistical_groupings    <list> "MEDC"
#> $ annotations.UN_M49.regions.AMERICAS         <chr> "019"
#> $ annotations.UN_M49.regions.NORTHERN_AMERICA <chr> "021"
#> $ annotations.UN_M49.regions.US               <chr> "840"
#> $ annotations.UN_M49.regions.WORLD            <chr> "001"
#> $ annotations.currency.alternate_symbols      <list> "US$"
#> $ annotations.currency.decimal_mark           <chr> "."
#> $ annotations.currency.disambiguate_symbol    <chr> "US$"
#> $ annotations.currency.html_entity            <chr> "$"
#> $ annotations.currency.iso_code               <chr> "USD"
#> $ annotations.currency.iso_numeric            <chr> "840"
#> $ annotations.currency.name                   <chr> "United States Dollar"
#> $ annotations.currency.smallest_denomination  <int> 1
#> $ annotations.currency.subunit                <chr> "Cent"
#> $ annotations.currency.subunit_to_unit        <int> 100
#> $ annotations.currency.symbol                 <chr> "$"
#> $ annotations.currency.symbol_first           <int> 1
#> $ annotations.currency.thousands_separator    <chr> ","
#> $ annotations.roadinfo.drive_on               <chr> "right"
#> $ annotations.roadinfo.road                   <chr> "South Miller Street"
#> $ annotations.roadinfo.speed_in               <chr> "mph"
#> $ annotations.sun.rise.apparent               <int> 1642434180
#> $ annotations.sun.rise.astronomical           <int> 1642427640
#> $ annotations.sun.rise.civil                  <int> 1642432140
#> $ annotations.sun.rise.nautical               <int> 1642429860
#> $ annotations.sun.set.apparent                <int> 1642380060
#> $ annotations.sun.set.astronomical            <int> 1642386540
#> $ annotations.sun.set.civil                   <int> 1642382100
#> $ annotations.sun.set.nautical                <int> 1642384380
#> $ annotations.timezone.name                   <chr> "America/Los_Angeles"
#> $ annotations.timezone.now_in_dst             <int> 0
#> $ annotations.timezone.offset_sec             <int> -28800
#> $ annotations.timezone.offset_string          <chr> "-0800"
#> $ annotations.timezone.short_name             <chr> "PST"
#> $ annotations.what3words.words                <chr> "mining.wires.secret"
#> $ bounds.northeast.lat                        <dbl> 47.42338
#> $ bounds.northeast.lng                        <dbl> -120.3248
#> $ bounds.southwest.lat                        <dbl> 47.42328
#> $ bounds.southwest.lng                        <dbl> -120.3249
#> $ `components.ISO_3166-1_alpha-2`             <chr> "US"
#> $ `components.ISO_3166-1_alpha-3`             <chr> "USA"
#> $ components._category                        <chr> "building"
#> $ components._type                            <chr> "building"
#> $ components.city                             <chr> "Wenatchee"
#> $ components.continent                        <chr> "North America"
#> $ components.country                          <chr> "United States"
#> $ components.country_code                     <chr> "us"
#> $ components.county                           <chr> "Chelan County"
#> $ components.house_number                     <chr> "18"
#> $ components.postcode                         <chr> "98801"
#> $ components.road                             <chr> "South Miller Street"
#> $ components.state                            <chr> "Washington"
#> $ components.state_code                       <chr> "WA"
#> $ geometry.lat                                <dbl> 47.42333
#> $ geometry.lng                                <dbl> -120.3249

# }