Get address data from the Photon API to OpenStreetMap

The Photon API is used get address data associated with the longitude and latitude. The following list of data is returned:

houseNumber

street

city

stateCode

stateName

zip

countryCode

countryName

The function makes an effort to convert both state and country Name into Code with codes defaulting to NA. Both Name and Code are returned so that improvements can be made in the conversion algorithm.

location_getSingleAddress_Photon(
  longitude = NULL,
  latitude = NULL,
  baseUrl = "https://photon.komoot.io/reverse",
  verbose = TRUE
)

Arguments

longitude
Single longitude in decimal degrees E.

latitude
Single latitude in decimal degrees N.

baseUrl
Base URL for data queries.

verbose
Logical controlling the generation of progress messages.

Value

List of address components.

References

https://photon.komoot.io

Examples

# \donttest{
library(MazamaLocationUtils)

# Set up standard directories and spatial data
spatialDataDir <- tempdir() # typically "~/Data/Spatial"
mazama_initialize(spatialDataDir)

# Wenatchee
lon <- -120.325278
lat <- 47.423333
addressList <- location_getSingleAddress_Photon(lon, lat)
str(addressList)

#> List of 8
#>  $ houseNumber: chr NA
#>  $ street     : chr NA
#>  $ city       : chr "Wenatchee"
#>  $ stateName  : chr "Washington"
#>  $ stateCode  : chr "WA"
#>  $ zip        : chr "98801"
#>  $ countryName: chr "United States"
#>  $ countryCode: chr "US"

# }