Create known location record with core metadata

Creates a known location record with the following columns of core metadata:

locationID

locationName

longitude

latitude

elevation

countryCode

stateCode

countyName

timezone

houseNumber

street

city

zip

location_initialize(
  longitude = NULL,
  latitude = NULL,
  stateDataset = "NaturalEarthAdm1",
  elevationService = NULL,
  addressService = NULL,
  verbose = TRUE
)

Arguments

longitude

Single longitude in decimal degrees E.

latitude

Single latitude in decimal degrees N.

stateDataset

Name of spatial dataset to use for determining state

elevationService

Name of the elevation service to use for determining the elevation. Default: NULL skips this step. Accepted values: "usgs".

addressService

Name of the address service to use for determining the street address. Default: NULL skips this step. Accepted values: "photon".

verbose

Logical controlling the generation of progress messages.

Value

Tibble with a single new known location.

Examples

# \donttest{
library(MazamaLocationUtils)

# Fail gracefully if any resources are not available
try({

  # Set up standard directories and spatial data
  spatialDataDir <- tempdir() # typically "~/Data/Spatial"
  mazama_initialize(spatialDataDir)

  # Wenatchee
  lon <- -120.325278
  lat <- 47.423333
  
  locationRecord <- location_initialize(lon, lat)
  str(locationRecord)
  
}, silent = FALSE)
#> tibble [1 × 13] (S3: tbl_df/tbl/data.frame)
#>  $ locationID  : chr "500cb09a70d3d981"
#>  $ locationName: chr "us.wa_500cb0"
#>  $ longitude   : num -120
#>  $ latitude    : num 47.4
#>  $ elevation   : logi NA
#>  $ countryCode : chr "US"
#>  $ stateCode   : chr "WA"
#>  $ countyName  : chr "Chelan"
#>  $ timezone    : chr "America/Los_Angeles"
#>  $ houseNumber : chr NA
#>  $ street      : chr NA
#>  $ city        : chr NA
#>  $ zip         : chr NA
# }