Return known locations near a target location

Returns a tibble of the known locations from locationTbl that are within distanceThreshold meters of the target location specified by longitude and latitude.

table_filterByDistance(
  locationTbl = NULL,
  longitude = NULL,
  latitude = NULL,
  distanceThreshold = NULL,
  measure = c("geodesic", "haversine", "vincenty", "cheap")
)

Arguments

locationTbl
Tibble of known locations.

longitude
Target longitude in decimal degrees E.

latitude
Target latitude in decimal degrees N.

distanceThreshold
Distance in meters.

measure
One of "haversine" "vincenty", "geodesic", or "cheap" specifying desired method of geodesic distance calculation.

Value

Tibble of known locations.

Note

Only a single target location is allowed.

Examples

library(MazamaLocationUtils)

locationTbl <- get(data("wa_monitors_500"))

# Too small a distanceThreshold will not find a match
locationTbl %>%
  table_filterByDistance(
    longitude = -117.3647, 
    latitude = 47.6725, 
    distanceThreshold = 10
  ) %>% 
  dplyr::glimpse()

#> Rows: 0
#> Columns: 13
#> $ locationID   <chr> 
#> $ locationName <chr> 
#> $ longitude    <dbl> 
#> $ latitude     <dbl> 
#> $ elevation    <dbl> 
#> $ countryCode  <chr> 
#> $ stateCode    <chr> 
#> $ countyName   <chr> 
#> $ timezone     <chr> 
#> $ houseNumber  <chr> 
#> $ street       <chr> 
#> $ city         <chr> 
#> $ zip          <chr> 

# Expanding the distanceThreshold will find several
locationTbl %>%
  table_filterByDistance(
    longitude = -117.3647, 
    latitude = 47.6725, 
    distanceThreshold = 10000
  ) %>%
  dplyr::glimpse()

#> Rows: 3
#> Columns: 13
#> $ locationID   <chr> "ced0aa4b3d45c5fc", "fa8288b1da3b2a87", "590324ebbb1c8896"
#> $ locationName <chr> "us.wa_ced0aa", "us.wa_fa8288", "us.wa_590324"
#> $ longitude    <dbl> -117.2576, -117.4263, -117.3649
#> $ latitude     <dbl> 47.66396, 47.69970, 47.67250
#> $ elevation    <dbl> 614.09, 624.69, 586.33
#> $ countryCode  <chr> "US", "US", "US"
#> $ stateCode    <chr> "WA", "WA", "WA"
#> $ countyName   <chr> "Spokane", "Spokane", "Spokane"
#> $ timezone     <chr> "America/Los_Angeles", "America/Los_Angeles", "America/Lo…
#> $ houseNumber  <chr> NA, NA, NA
#> $ street       <chr> NA, NA, NA
#> $ city         <chr> "Spokane Valley", "Spokane", "Spokane"
#> $ zip          <chr> "99206", "99205", "99207"