Finds overlapping locations in a known locations table.

Calculates distances between all locations within a known locations table and returns a tibble with the row indices and separation distances of those records with overlapping locations.

It is useful when working with new metadata tables to identify overlapping locations early on so that decisions can be made about the apporpriateness of the specified radius.

table_findOverlappingLocations(tbl = NULL, radius = NULL)

Arguments

tbl
Tibble with longitude and latitude columns.

radius
Radius in meters.

Value

Tibble of row indices and distances for those locations which overlap.

Examples

library(MazamaLocationUtils)

meta <- wa_airfire_meta

# Anything locations closer than 2 km? (diameter = 2*radius)
table_findOverlappingLocations(meta, radius = 1000)

#> # A tibble: 0 × 3
#> # … with 3 variables: row1 <dbl>, row2 <dbl>, distance <dbl>

# How about 4 km?
table_findOverlappingLocations(meta, radius = 2000)

#> # A tibble: 3 × 3
#>    row1  row2 distance
#>   <int> <int>    <dbl>
#> 1     2    34    2497.
#> 2    12    34    3329.
#> 3     2     9    3762.

# Let's look at those locations

tooCloseTbl <- table_findOverlappingLocations(meta, radius = 2000)

for ( i in seq_len(nrow(tooCloseTbl)) ) {
  rows <- as.numeric(tooCloseTbl[i, 1:2])
  cat(sprintf("\n%5.1f meters apart:\n", tooCloseTbl$distance[i]))
  print(meta[rows, c('longitude', 'latitude', 'siteName')])
}

#> 
#> 2496.6 meters apart:
#>              longitude latitude            siteName
#> 530330057_01 -122.3405 47.56200    Seattle-Duwamish
#> 530330080_01 -122.3086 47.56824 Seattle-Beacon Hill
#> 
#> 3329.0 meters apart:
#>              longitude latitude              siteName
#> 530330030_01 -122.3197 47.59722 Seattle-10th & Weller
#> 530330080_01 -122.3086 47.56824   Seattle-Beacon Hill
#> 
#> 3762.3 meters apart:
#>              longitude latitude           siteName
#> 530330057_01 -122.3405  47.5620   Seattle-Duwamish
#> 530331011_01 -122.3208  47.5309 Seattle-South Park