Create an empty known location table

Creates an empty known location tibble with the following columns of core metadata:

locationID

locationName

longitude

latitude

elevation

countryCode

stateCode

countyName

timezone

houseNumber

street

city

zip

table_initialize()

Value

Empty known location tibble with the specified metadata columns.

Examples

library(MazamaLocationUtils)

# Create an empty Tbl
emptyTbl <- table_initialize()
dplyr::glimpse(emptyTbl)

#> Rows: 0
#> Columns: 13
#> $ locationID   <chr> 
#> $ locationName <chr> 
#> $ longitude    <dbl> 
#> $ latitude     <dbl> 
#> $ elevation    <dbl> 
#> $ countryCode  <chr> 
#> $ stateCode    <chr> 
#> $ countyName   <chr> 
#> $ timezone     <chr> 
#> $ houseNumber  <chr> 
#> $ street       <chr> 
#> $ city         <chr> 
#> $ zip          <chr>