Load a known location table

Load a tibble of known locations from the preferred directory.

The known location table must be named either <collectionName>.rda or <collectionName>.csv. If both are found, only <collectionName>.rda will be loaded to ensure that columns will have the proper type assigned.

table_load(collectionName = NULL)

Arguments

collectionName

Character identifier for this table.

Value

Tibble of known locations.

See also

Examples

library(MazamaLocationUtils)

# Set the directory for saving location tables
setLocationDataDir(tempdir())

# Load an example table and check the dimensions
locationTbl <- get(data("wa_monitors_500"))
dim(locationTbl)
#> [1] 78 13

# Save it as "table_load_example"
table_save(locationTbl, "table_load_example")

# Load it and check the dimensions
my_table <- table_load("table_load_example")
dim(my_table)
#> [1] 78 13

# Check the locationDataDir
list.files(getLocationDataDir(), pattern = "table_load_example")
#> [1] "table_load_example.rda"