Save a known location table

Save a tibble of known locations to the preferred directory.

table_save(
  locationTbl = NULL,
  collectionName = NULL,
  backup = TRUE,
  outputType = c("rda", "csv")
)

Arguments

locationTbl
Tibble of known locations.

collectionName
Character identifier for this table.

backup
Logical specifying whether to save a backup version of any existing tables sharing collectionName.

outputType
Output format. One of "rda" or "csv".

Value

File path of saved file.

Details

Backup files are saved with "YYYY-mm-ddTHH:MM:SS"

Examples

library(MazamaLocationUtils)

# Set the directory for saving location tables
setLocationDataDir(tempdir())

# Load an example table and check the dimensions
locationTbl <- get(data("wa_monitors_500"))
dim(locationTbl)

#> [1] 68 13

# Save it as "table_save_example"
table_save(locationTbl, "table_save_example")

# Add a column and save again
locationTbl %>% 
  table_addColumn("my_column") %>% 
  table_save("table_save_example")
  
# Check the locationDataDir
list.files(getLocationDataDir(), pattern = "table_save_example")

#> [1] "table_save_example.2022-01-05T16:35:26.rda"
#> [2] "table_save_example.rda"