Create known location record with core metadata

Creates a known location record with the following columns of core metadata:

locationID

locationName

longitude

latitude

elevation

countryCode

stateCode

countyName

timezone

houseNumber

street

city

zip

location_initialize(
  longitude = NULL,
  latitude = NULL,
  stateDataset = "NaturalEarthAdm1",
  elevationService = NULL,
  addressService = NULL,
  verbose = TRUE
)

Arguments

longitude
Single longitude in decimal degrees E.

latitude
Single latitude in decimal degrees N.

stateDataset
Name of spatial dataset to use for determining state

elevationService
Name of the elevation service to use for determining the elevation. Default: NULL. Accepted values: "usgs".

addressService
Name of the address service to use for determining the street address. Default: NULL Accepted values: "photon".

verbose
Logical controlling the generation of progress messages.

Value

Tibble with a single new known location.

Examples

# \donttest{
library(MazamaLocationUtils)

# Set up standard directories and spatial data
spatialDataDir <- tempdir() # typically "~/Data/Spatial"
mazama_initialize(spatialDataDir)

# Wenatchee
lon <- -120.325278
lat <- 47.423333
locationRecord <- location_initialize(lon, lat)
# }