Finds adjacent locations in a known locations table.

Calculate distances between all locations within a known locations table and return a tibble containing all records that have an adjacent location separated by less than distanceThreshold meters. The return tibble is ordered by separation distance.

It is useful when working with new metadata tables to identify adjacent locations early on so that decisions can be made about the appropriateness of the specified distanceThreshold.

table_findAdjacentLocations(
  locationTbl = NULL,
  distanceThreshold = NULL,
  measure = c("geodesic", "haversine", "vincenty", "cheap")
)

Arguments

locationTbl
Tibble of known locations.

distanceThreshold
Distance in meters.

measure
One of "haversine" "vincenty", "geodesic", or "cheap" specifying desired method of geodesic distance calculation.

See geodist::geodist for details.

Value

Tibble of known locations separated by less than distanceThreshold meters.

Note

The measure "cheap" may be used to speed things up depending on the spatial scale being considered. Distances calculated with measure = "cheap" will vary by a few meters compared with those calculated using measure = "geodesic".

Examples

library(MazamaLocationUtils)

meta <- wa_airfire_meta

# Any locations closer than 2 km?
meta %>%
  table_findAdjacentLocations(distanceThreshold = 2000) %>%
  dplyr::select(siteName, timezone)

#>                                   siteName            timezone
#> lon_.122.671_lat_47.029_usfs.1078          America/Los_Angeles
#> lon_.122.677_lat_47.028_usfs.1078          America/Los_Angeles

# How about 4 km?
meta %>%
  table_findAdjacentLocations(distanceThreshold = 4000) %>%
  dplyr::select(siteName, timezone)

#>                                                siteName            timezone
#> lon_.122.671_lat_47.029_usfs.1078                       America/Los_Angeles
#> lon_.122.677_lat_47.028_usfs.1078                       America/Los_Angeles
#> 530330057_01                           Seattle-Duwamish America/Los_Angeles
#> 530330080_01                        Seattle-Beacon Hill America/Los_Angeles
#> 530330030_01                      Seattle-10th & Weller America/Los_Angeles
#> 530331011_01                         Seattle-South Park America/Los_Angeles