Create an empty known location table

Creates an empty known location tibble with the following columns of core metadata:

locationID

locationName

longitude

latitude

elevation

countryCode

stateCode

county

timezone

houseNumber

street

city

zip

table_initialize()

Value

Empty known location tibble with the specified metadata columns.

Examples

if (FALSE) {
# Set up standard directories and spatial data
mazama_initialize()

# Create an empty Tbl
emptyTbl <- table_initialize()
}