Update a column of metadata in a table

For matching locationID records the assaociated locatioData is used to replace any existing value in columnName.

table_updateColumn(locationTbl = NULL, columnName = NULL,
  locationID = NULL, locationData = NULL, verbose = TRUE)

Arguments

locationTbl
Tibble of known locations, Default: NULL

columnName
Name to use for the new column, Default: NULL

locationID
Vector of locationID strings, Default: NULL

locationData
Vector of data to used at matching records, Default: NULL

verbose
Logical controlling the generation of progress messages.

Value

Updated tibble of known locations.

See also

Examples

locationTbl <- get(data("wa_monitors_500"))
wa <- get(data("wa_airfire_meta"))

# We will merge some metadata from wa into locationTbl

# Record indices for wa
wa_indices <- seq(5,65,5)
wa_sub <- wa[wa_indices,]

locationID <- table_getLocationID(locationTbl, wa_sub$longitude, wa_sub$latitude, radius = 500)
locationData <- wa_sub$siteName

locationTbl <- table_updateColumn(locationTbl, "siteName", locationID, locationData)

# Look at the data we attempted to merge
wa$siteName[wa_indices]

#>  [1] "Chelan-Woodin Ave"            "Lacrosse-Hill St"            
#>  [3] "Kennewick-Metaline"           "Sunnyside-S 16th"            
#>  [5] "Inchelium"                    "Wellpinit-Spokane Tribe"     
#>  [7] "Lake Forest Park-Town Center" "Twisp-Glover St"             
#>  [9] "Maple Falls-Azure Way"        "Omak-Colville Tribe"         
#> [11] "Ritzville-Alder St "          "Darrington-Fir St"           
#> [13] "Tukwila_Allentown"           

# And two columns from the updated locationTbl
locationTbl_indices <- table_getRecordIndex(locationTbl, locationID)
locationTbl[locationTbl_indices, c("city","siteName")]

#> # A tibble: 13 x 2
#>    city               siteName                    
#>    <chr>              <chr>                       
#>  1 Chelan             Chelan-Woodin Ave           
#>  2 La Crosse          Lacrosse-Hill St            
#>  3 Tri-Cities         Kennewick-Metaline          
#>  4 Sunnyside          Sunnyside-S 16th            
#>  5 Inchelium          Inchelium                   
#>  6 Wellpinit          Wellpinit-Spokane Tribe     
#>  7 Lake Forest Park   Lake Forest Park-Town Center
#>  8 Okanogan County    Twisp-Glover St             
#>  9 Limestone Junction Maple Falls-Azure Way       
#> 10 Okanogan County    Omak-Colville Tribe         
#> 11 Ritzville          "Ritzville-Alder St "       
#> 12 Darrington         Darrington-Fir St           
#> 13 Tukwila            Tukwila_Allentown